requirements.txt
Django==5.0.6
django-ckeditor==6.7.1
djangorestframework==3.15.1
pillow==10.3.0
psycopg2==2.9.9
gunicorn==20.1.0
argon2-cffi==20.1.0